requirements.txt
huggingface_hub
torch
torchvision
Pillow
scikit-image
numpy
streamlit
fastai